redesign scripts
