--- .flake8
@@ -1,5 +1,0 @@
-[flake8]
-ignore = E501, W503, F403, F401, C901, F541
-max-line-length = 120
-max-complexity = 18
-select = B,C,E,F,W,T4,B9
