+++ .pre-commit-config.yaml
@@ -6,14 +6,11 @@ repos:
       - id: end-of-file-fixer
       - id: check-yaml
       - id: requirements-txt-fixer
-  - repo: https://github.com/pre-commit/mirrors-autopep8
-    rev: v1.6.0
+  - repo: https://github.com/psf/black
+    rev: 22.10.0
     hooks:
-      - id: autopep8
-  - repo: https://github.com/PyCQA/flake8
-    rev: 4.0.1
-    hooks:
-      - id: flake8
+    -   id: black
+        args: [--line-length=120]
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v0.950
     hooks:
@@ -24,4 +21,4 @@ repos:
     hooks:
       - id: isort
         name: isort (python)
-        args: ["--force-single-line-imports"]
+        args: ["--profile google"]
