--- configs/hydra/job_logging/custom.yaml
@@ -1,8 +1,0 @@
-version: 1
-handlers:
-  console:
-    class: rich.logging.RichHandler
-root:
-  handlers: [console]
-
-disable_existing_loggers: false
