--- configs/method/simhash.yaml
@@ -1,23 +1,0 @@
-dataset: oscar-corpus/OSCAR-2109
-configs:
-  - deduplicated_gl
-columns:
-  - text
-num_proc: 10
-num_samples: -1
-embedder:
-  name: SimHashEmbedder
-  hamming_distance: 3
-  num_blocks: 4
-tokenization:
-  ngram_size: 2
-  level: word
-cache_dir: .cache
-storage_config:
-  type: redis
-  redis:
-    host: localhost
-    port: 6379
-# storage_config: null # use null to use dict instead
-output: simhash_results.json
-verbose: true
