--- configs/method/suffix.yaml
@@ -1,15 +1,0 @@
-dataset: oscar-corpus/OSCAR-2109
-configs:
-  - deduplicated_gl
-columns:
-  - text
-num_proc: 10
-num_samples: -1
-embedder:
-  name: SuffixArrayEmbedder
-  k: 100
-  cache_dir: outputs
-  temp_file_prefix: sa
-cache_dir: .cache
-output: sa_results.json
-verbose: true
