+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "text-dedup"
-version = "0.2.1"
+version = "0.3.0"
 description = "All-in-one text deduplication"
 authors = ["Chenghao Mou <mouchenghao@gmail.com>"]
 license = "Apache-2.0"
@@ -11,23 +11,12 @@ packages = [{include = "text_dedup"}]
 python = "^3.9,<3.12"
 numpy = "^1.23.2"
 datasketch = "^1.5.8"
-torch = "^1.12.1"
-annoy = "^1.17.1"
 tqdm = "^4.64.1"
-hydra-core = "^1.2.0"
-transformers = "^4.21.2"
 datasets = "^2.4.0"
-sentencepiece = "^0.1.97"
-redis = "^4.3.4"
 rich = "^12.5.1"
-mpire = "^2.6.0"
 scipy = "1.9.1"
 xxhash = "^3.0.0"
-sentence-transformers = "^2.2.2"
-dpu-utils = "^0.6.1"
-typer = "^0.6.1"
-pybloom = {git = "https://github.com/jaybaird/python-bloomfilter"}
-networkx = "^2.8.7"
+pybloom-live = "^3.1.1"
 networkit = "^10.0"
 
 [tool.poetry.group.dev.dependencies]
