--- text_dedup/exact_dedup/suffix_array/__init__.py
@@ -1,6 +1,0 @@
-from text_dedup.exact_dedup.suffix_array.suffix_array_google import \
-    GoogleSuffixArray
-from text_dedup.exact_dedup.suffix_array.suffix_array_python import \
-    PythonSuffixArray
-
-__all__ = ['GoogleSuffixArray', 'PythonSuffixArray']
