+++ text_dedup/exact_hash.py
@@ -1,0 +1,82 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# @Date    : 2022-11-05 09:44:48
+# @Author  : Chenghao Mou (mouchenghao@gmail.com)
+import argparse
+from hashlib import md5
+import logging
+import os
+import time
+
+from datasets import load_dataset
+from rich.logging import RichHandler
+
+from text_dedup.utils import add_exact_hash_args
+from text_dedup.utils import add_io_args
+from text_dedup.utils import add_meta_args
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+logger.addHandler(RichHandler(rich_tracebacks=True))
+logger.propagate = False
+
+if __name__ == "__main__":
+
+    parser = argparse.ArgumentParser(
+        prog="text_dedup.exacthash",
+        description="Deduplicate text using exact hashing",
+    )
+    parser = add_io_args(parser)
+    parser = add_meta_args(parser)
+    parser = add_exact_hash_args(parser)
+    args = parser.parse_args()
+
+    if args.path is not None:
+        elapsed_time = {"All": time.time()}
+        # region: loading
+        elapsed_time["Loading"] = time.time()
+        ds = load_dataset(
+            path=args.path,
+            name=args.name,
+            data_dir=args.data_dir,
+            data_files=args.data_files,
+            split=args.split,
+            cache_dir=args.cache_dir,
+            use_auth_token=args.use_auth_token,
+        )
+        elapsed_time["Loading"] = time.time() - elapsed_time["Loading"]
+        # endregion
+
+        hash_func = {
+            "md5": md5,
+        }[args.hash_func]
+        hashes = set()
+        flags = []
+
+        # region: processing
+        elapsed_time["Processing"] = time.time()
+        for example in ds[args.column]:
+            h = hash_func(example.encode("utf-8")).hexdigest()
+            if h in hashes:
+                flags.append(True)
+            else:
+                flags.append(False)
+                hashes.add(h)
+        elapsed_time["Processing"] = time.time() - elapsed_time["Processing"]
+        # endregion
+
+        # region: filtering and save
+        elapsed_time["Filtering"] = time.time()
+        ds = ds.filter(lambda _, idx: not flags[idx], with_indices=True, num_proc=os.cpu_count())
+        elapsed_time["Filtering"] = time.time() - elapsed_time["Filtering"]
+        elapsed_time["Saving"] = time.time()
+        ds.save_to_disk(os.path.join(args.output_dir, args.dedup_name))
+        elapsed_time["Saving"] = time.time() - elapsed_time["Saving"]
+        elapsed_time["All"] = time.time() - elapsed_time["All"]
+        # endregion
+
+        for k, v in elapsed_time.items():
+            logger.info(f"{k:<30}: {v:.2f}s")
+
+        logger.info(f"{'Before':<30}: {len(flags)}")
+        logger.info(f"{'After':<30}: {len(ds)}")
