--- text_dedup/near_dedup/__init__.py
@@ -1,5 +1,0 @@
-from text_dedup.near_dedup.minhash import MinHashEmbedder
-from text_dedup.near_dedup.simhash import SimHashEmbedder
-from text_dedup.near_dedup.simhash import SimHashIndex
-
-__all__ = ['SimHashEmbedder', 'SimHashIndex', 'MinHashEmbedder']
