--- text_dedup/near_dedup/minhash/__init__.py
@@ -1,5 +1,0 @@
-from text_dedup.near_dedup.minhash.minhash_deduplicator import \
-    MinHashDeduplicator
-from text_dedup.near_dedup.minhash.minhash_embedder import MinHashEmbedder
-
-__all__ = ['MinHashEmbedder', 'MinHashDeduplicator']
