--- text_dedup/near_dedup/simhash/__init__.py
@@ -1,6 +1,0 @@
-from text_dedup.near_dedup.simhash.simhash_deduplicator import \
-    SimHashDeduplicator
-from text_dedup.near_dedup.simhash.simhash_embedder import SimHashEmbedder
-from text_dedup.near_dedup.simhash.simhash_index import SimHashIndex
-
-__all__ = ['SimHashEmbedder', 'SimHashIndex', 'SimHashDeduplicator']
