--- text_dedup/postprocess/__init__.py
@@ -1,7 +1,0 @@
-from text_dedup.postprocess.clustering import annoy_clustering
-from text_dedup.postprocess.clustering import lsh_clustering
-from text_dedup.postprocess.clustering import simhash_clustering
-from text_dedup.postprocess.group import get_group_indices
-
-__all__ = ['annoy_clustering', 'lsh_clustering',
-           'simhash_clustering', 'get_group_indices']
