--- text_dedup/preprocess/__init__.py
@@ -1,3 +1,0 @@
-from text_dedup.preprocess.tokenization import tokenize
-
-__all__ = ['tokenize']
