--- text_dedup/semantic_dedup/__init__.py
@@ -1,3 +1,0 @@
-from text_dedup.semantic_dedup.transformer import TransformerEmbedder
-
-__all__ = ['TransformerEmbedder']
