+++ text_dedup/utils.py
@@ -1,0 +1,73 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# @Date    : 2022-11-05 09:16:34
+# @Author  : Chenghao Mou (mouchenghao@gmail.com)
+import argparse
+
+
+def add_io_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
+    parser.add_argument("--path", type=str, help="Path to use in load_dataset", required=True),
+    parser.add_argument("--name", type=str, help="Dataset name to use in load_dataset"),
+    parser.add_argument("--data_dir", type=str, help="Data directory to use in load_dataset"),
+    parser.add_argument("--data_files", type=str, help="Data files to use in load_dataset"),
+    parser.add_argument("--split", type=str, help="Split to use in load_dataset"),
+    parser.add_argument("--cache_dir", type=str, help="Cache directory to use in load_dataset"),
+    parser.add_argument(
+        "--use_auth_token", action=argparse.BooleanOptionalAction, help="Use auth token to use in load_dataset"
+    ),
+    parser.add_argument("--output_dir", type=str, help="Path to save outputs"),
+    parser.add_argument("--index_name", type=str, help="Name of the index file, will be saved in save_dir"),
+    parser.add_argument("--reuse_index", action=argparse.BooleanOptionalAction, help="Reuse the index if it exists"),
+    parser.add_argument("--graph_name", type=str, help="Name of the cluster file, will be saved in save_dir"),
+    parser.add_argument("--reuse_graph", action=argparse.BooleanOptionalAction, help="Reuse the cluster if it exists"),
+    parser.add_argument("--dedup_name", type=str, help="Name of the deduplicated dataset, will be saved in save_dir"),
+    return parser
+
+
+def add_meta_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
+    parser.add_argument("--column", type=str, help="Single column to use for deduplication", required=True),
+    return parser
+
+
+def add_minhash_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
+    parser.add_argument("--ngram", type=int, default=3, help="Ngram to use in MinHash"),
+    parser.add_argument("--seed", type=int, default=42, help="Seed to use in MinHash"),
+    parser.add_argument("--num_perm", type=int, default=128, help="Number of permutations to use in MinHash"),
+    parser.add_argument("--threshold", type=float, default=0.8, help="Threshold to use in MinHashLSH"),
+    return parser
+
+
+def add_simhash_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
+    parser.add_argument("--ngram", type=int, default=3, help="Ngram to use in SimHash"),
+    parser.add_argument("--seed", type=int, default=42, help="Seed to use in SimHash"),
+    parser.add_argument("--bit_diff", type=int, default=3, help="Bit difference to use in SimHash"),
+    parser.add_argument(
+        "--num_bucket", type=int, default=4, help="Number of buckets to use in SimHash, must be larger than bit_diff"
+    ),
+    return parser
+
+
+def add_sa_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
+    parser.add_argument(
+        "--k", type=int, default=100, help="Minimum byte length of a duplicate substring in Suffix Array Deduplication"
+    ),
+    parser.add_argument(
+        "--strategy",
+        type=str,
+        default="overlapping",
+        help="Strategy when there are overlapping duplicate substrings",
+        choices=["overlapping", "longest"],
+    )
+    parser.add_argument("--google_repo_path", type=str, help="Path to google-research-deduplication", required=True),
+    return parser
+
+
+def add_bloom_filter_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
+    parser.add_argument("--error_rate", type=float, default=1e-6, help="Error rate to use in BloomFilter"),
+    parser.add_argument("--hash_func", type=str, default="md5", help="Hash function to use in BloomFilter"),
+    return parser
+
+
+def add_exact_hash_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
+    parser.add_argument("--hash_func", type=str, default="md5", help="Hash function to use in ExactHash"),
+    return parser
