--- text_dedup/utils/hf_datasets/__init__.py
@@ -1,13 +1,0 @@
-from text_dedup.utils.hf_datasets.helpers import dataset_get
-from text_dedup.utils.hf_datasets.helpers import dataset_get_all_str_columns
-from text_dedup.utils.hf_datasets.helpers import dataset_map
-from text_dedup.utils.hf_datasets.helpers import extract_text
-from text_dedup.utils.hf_datasets.helpers import get_byte_size
-
-__all__ = [
-    "get_byte_size",
-    "extract_text",
-    "dataset_map",
-    "dataset_get",
-    "dataset_get_all_str_columns",
-]
