--- text_dedup/utils/similarity/__init__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# author      : Chenghao Mou (mouchenghao@gmail.com)
-# created     : 9/29/22
-# description : {}
-
-"""
-A longer description that spans multiple lines.  Explain the purpose of the
-file and provide a short list of the key classes/functions it contains.
-"""
